# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dispatcher API",
    "version": "1.0.0",
    "description": "Backend for the Dispatcher Expo app — auth, onboarding, drives board, balances."
  },
  "servers": [{ "url": "https://api-production-f4ac.up.railway.app" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/v1/favorites": {
      "get": {
        "security": [{ "bearerAuth": [] }],
        "summary": "List users the current user has favorited",
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "security": [{ "bearerAuth": [] }],
        "summary": "Favorite a user (dispatcher↔driver). Body: { userId }",
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/v1/favorites/{userId}": {
      "delete": {
        "security": [{ "bearerAuth": [] }],
        "summary": "Unfavorite a user",
        "responses": { "204": { "description": "No content" } }
      }
    }
  }
}